Extracted invoice printing to separate class
